Initial commit for refactoring
