+++ src/js/components/core/create_project/ProjectModal.js
@@ -8,15 +8,15 @@ const Button = require('react-bootstrap/lib/Button.js');
 const ButtonGroup = require('react-bootstrap/lib/ButtonGroup.js');
 const ButtonToolbar = require('react-bootstrap/lib/ButtonToolbar.js');
 const Checkbox = require('react-bootstrap/lib/Checkbox.js');
-const CoreStore = require('../../stores/CoreStore.js');
-const CheckStore = require('../.././stores/CheckStore');
-const CoreActions = require('../../actions/CoreActions.js');
+const CoreStore = require('../../../stores/CoreStore.js');
+const CheckStore = require('../../../stores/CheckStore');
+const CoreActions = require('../../../actions/CoreActions.js');
 const CheckDataGrabber = require('./CheckDataGrabber');
 const {dialog} = window.electron.remote;
-const Loader = require('./Loader');
+const Loader = require('../Loader');
 const ENTER = 13;
-const booksOfBible = require('./booksOfBible');
-const TargetLanguage = require('./UploadModal');
+const booksOfBible = require('../booksOfBible');
+const TargetLanguage = require('../UploadModal');
 const SelectChecks = require('./SelectChecks');
 const ProjectName = require('./ProjectName');
