+++ src/js/pages/app.js
@@ -9,7 +9,7 @@ const NavBarComponent = require('../components/core/NavBarComponent');
 const LoginModal = require('../components/core/LoginModal');
 const SettingsModal = require('../components/core/SettingsModal.js');
 // const RootStyles = require('./RootStyle');
-const ProjectModal = require('../components/core/ProjectModal');
+const ProjectModal = require('../components/core/create_project/ProjectModal');
 const Loader = require('../components/core/Loader');
 const RootStyles = require('./RootStyle');
 const Grid = require('react-bootstrap/lib/Grid.js');
